# openapi.json
{
  "components": {
    "schemas": {
      "Configuration": {
        "properties": {
          "forwards": {
            "items": {
              "$ref": "#/components/schemas/Forward"
            },
            "type": "array"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/Group"
            },
            "type": "array"
          },
          "peers": {
            "items": {
              "$ref": "#/components/schemas/Peer"
            },
            "type": "array"
          },
          "revision": {
            "format": "int64",
            "type": "integer"
          },
          "settings": {
            "allOf": [
              {
                "$ref": "#/components/schemas/NetworkSettings"
              }
            ],
            "nullable": true
          }
        },
        "required": [
          "revision",
          "groups",
          "peers",
          "forwards"
        ],
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "persisted_revision": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "Forward": {
        "properties": {
          "allowed_group_ids": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "protocol": {
            "type": "string"
          },
          "target_peer_id": {
            "type": "string"
          },
          "target_port": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "protocol",
          "target_peer_id",
          "target_port",
          "allowed_group_ids"
        ],
        "type": "object"
      },
      "Group": {
        "properties": {
          "allowed_groups": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "type": "object"
      },
...
      "NetworkSettings": {
        "properties": {
          "endpoint": {
            "type": "string"
          },
          "persistent_keepalive": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "subnet": {
            "type": "string"
          }
        },
        "required": [
          "subnet",
          "endpoint",
          "persistent_keepalive"
        ],
        "type": "object"
      },
      "NewForward": {
        "additionalProperties": false,
        "properties": {
          "allowed_group_ids": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "name": {
            "type": "string"
          },
          "protocol": {
            "type": "string"
          },
          "target_peer_id": {
            "type": "string"
          },
          "target_port": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "name",
          "protocol",
          "target_peer_id",
          "target_port",
          "allowed_group_ids"
        ],
        "type": "object"
      },
...
      "Peer": {
        "properties": {
          "group_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "ipv4": {
            "type": "string"
          },
          "last_handshake_unix": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "public_key": {
            "type": "string"
          },
          "received_bytes": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "sent_bytes": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "public_key",
          "ipv4",
          "group_id",
          "received_bytes",
          "sent_bytes"
        ],
        "type": "object"
      },
...
      "PolicyChange": {
        "additionalProperties": false,
        "properties": {
          "allowed_groups": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "group_id": {
            "type": "string"
          }
        },
        "required": [
          "group_id",
          "allowed_groups"
        ],
        "type": "object"
      },
      "PolicyRequest": {
        "additionalProperties": false,
        "properties": {
          "changes": {
            "items": {
              "$ref": "#/components/schemas/PolicyChange"
            },
            "type": "array"
          },
          "expected_revision": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision",
          "changes"
        ],
        "type": "object"
      },
      "PolicyResult": {
        "properties": {
          "applied_revision": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/Group"
            },
            "type": "array"
          },
          "revision": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "revision",
          "groups"
        ],
        "type": "object"
      },
      "RuntimeStatus": {
        "properties": {
          "applied_revision": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "last_activation_error": {
            "nullable": true,
            "type": "string"
          },
          "persisted_revision": {
            "format": "int64",
            "type": "integer"
          },
          "ready": {
            "type": "boolean"
          }
        },
        "required": [
          "ready",
          "persisted_revision"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "adminBearer": {
        "description": "Exact configured WIREHUB_ADMIN_TOKEN. Keep it outside URLs and logs.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "",
    "license": {
      "name": ""
    },
    "title": "wirehub",
    "version": "1.0.0-rc.1"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/config": {
      "get": {
        "operationId": "get_config",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Configuration"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Persistence failure"
          }
        },
        "security": [
          {
            "adminBearer": []
          }
        ],
        "tags": [
          "configuration"
        ]
      }
    },
    "/api/forwards": {
      "get": {
        "operationId": "list_forwards",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Forward"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Persistence failure"
          }
        },
        "security": [
          {
            "adminBearer": []
          }
        ],
        "tags": [
          "crate"
        ]
      },
      "post": {
        "operationId": "create_forward",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewForward"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Forward"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Body too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "JSON content type required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid JSON body"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Persistence failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Runtime unavailable or activation unconfirmed"
          }
        },
        "security": [
          {
            "adminBearer": []
          }
        ],
        "tags": [
          "crate"
        ]
      }
    },
    "/api/forwards/{id}": {
      "delete": {
        "operationId": "delete_forward",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource missing"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Persistence failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Runtime unavailable or activation unconfirmed"
          }
        },
        "security": [
          {
            "adminBearer": []
          }
        ],
        "tags": [
          "crate"
        ]
      }
    },
...
    "/api/policy": {
      "put": {
        "operationId": "put_policy",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyResult"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Body too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "JSON content type required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid JSON body"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Persistence failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Runtime unavailable or activation unconfirmed"
          }
        },
        "security": [
          {
            "adminBearer": []
          }
        ],
        "tags": [
          "configuration"
        ]
      }
    },
...
    "/api/status": {
      "get": {
        "operationId": "get_status",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeStatus"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Persistence failure"
          }
        },
        "security": [
          {
            "adminBearer": []
          }
        ],
        "tags": [
          "configuration"
        ]
      }
    }
  }
}